+++ app/src/main/java/com/alorma/notifix/domain/model/CreateAppNotification.kt
@@ -1,4 +1,4 @@
 package com.alorma.notifix.domain.model
 
 data class CreateAppNotification(val title: String, val text: String?, val color: Int,
-                                 val checked: Boolean, val triggerId: Long?)
+                                 val checked: Boolean, val triggerId: Int?)
