+++ app/src/main/java/com/alorma/notifix/ui/features/create/CreateNotificationAction.kt
@@ -17,4 +17,4 @@ data class PreviewNotificationAction(val title: String,
                                      val checked: Boolean,
                                      val color: Int) : CreateNotificationAction()
 
-class TriggerCreatedAction(val triggerId: Long): CreateNotificationAction()
+class TriggerCreatedAction(val triggerId: Int): CreateNotificationAction()
